Move deserialization stuff to a new file.
